Parametric experiment for italic (note problem)
Somehow the YTLX and YTFI axes are getting linked so that the YTLX axis below zero has no effect unless YTFI is set to its max. You can see the problem in a TTX of the gvar table, where a tuple for (e.g.) "a" has coord elements for both YTLX and YTFI. Delete the YTFI coord and it works again. This happens with fontmake but not when the font is exported directly from Glyphs.